// src/app/Model/Contabil.ts
export class Contabil {

     idUsuario: Number = 0;
     descricao: String = "";
     transporte: String = "0.0";
     garantiaExtra: String = "0.0";
     icms: String = "0.0";
     pis: String = "0.0";
     cofins: String = "0.0";
     ipi: String = "0.0";
     iss: String ="0.0";
     margem: String = "0.0";
}